pyproject.toml
[project]
name = "advanced-rag"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "agno>=2.3.3",
    "agno-os>=0.1.0",
    "deepeval>=3.7.4",
    "fastapi>=0.122.0",
    "fastembed>=0.7.3",
    "google-genai>=1.53.0",
    "hyperopt>=0.2.7",
    "langchain-community>=0.4.1",
    "langchain-mistralai>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "llama-cpp-python>=0.3.16",
    "openai>=2.8.1",
    "qdrant-client>=1.16.1",
    "ragas>=0.3.9",
    "rapidfuzz>=3.14.3",
    "sentence-transformers>=5.1.2",
    "uvicorn>=0.38.0",
    "wandb>=0.23.1",
]
